+++ pandapower/toolbox/grid_modification.py
@@ -1813,7 +1813,8 @@ def replace_ward_by_internal_elements(net: pandapowerNet, wards: Collection[int]
                                   * net.ward.qz_mvar.loc[wards].to_numpy()
                                   * sign_in_service
                                   * sign_not_isolated)
-        df_lst: list[pd.DataFrame] = [net.res_shunt, to_add_shunt]
+
+        df_lst = [net.res_shunt, to_add_shunt]
         net.res_shunt = pd.concat(df_lst)
 
     # --- drop replaced wards